+++ streamlabs/custom.js
@@ -1,5 +1,8 @@
 const settings = {
-  "showPronouns": true,
+  "pronouns": [
+    "api.pronouns.alejo.io",
+    "pronoundb.org"
+  ],
   "showFrogEmotes": true,
   "capitalizePronouns": true
 };
@@ -17,9 +20,8 @@ const overrides = {
     // "emote_id": null, // remove emote
   }
 };
-const pronounsApis = ["api.pronouns.alejo.io","pronoundb.org"];
-const userAgent = "pronouns-chat/5.0.0 (https://github.com/liquidnya/pronouns-chat)";
-/* version 5.0.0 */
+const userAgent = "pronouns-chat/6.0.0 (https://github.com/liquidnya/pronouns-chat)";
+/* version 6.0.0 */
 "use strict";
 (() => {
   var __create = Object.create;
@@ -5916,7 +5918,7 @@ const userAgent = "pronouns-chat/5.0.0 (https://github.com/liquidnya/pronouns-ch
       }
     },
     async load(api) {
-      if (api.settings.showPronouns) {
+      if (api.settings.pronouns.includes("api.pronouns.alejo.io")) {
         this.map = null;
         this.cache = Caches.builder().expireAfterWrite(Time.minutes(5)).buildAsync((key) => this.fetchPronouns(key));
         void this.loadPronounsMap();
@@ -6057,7 +6059,7 @@ const userAgent = "pronouns-chat/5.0.0 (https://github.com/liquidnya/pronouns-ch
       }
     },
     async load(api) {
-      if (api.settings.showPronouns) {
+      if (api.settings.pronouns.includes("pronoundb.org")) {
         const batch = createBatch(
           (ids) => this.fetchPronounsBatch(ids),
           50
@@ -6092,9 +6094,9 @@ const userAgent = "pronouns-chat/5.0.0 (https://github.com/liquidnya/pronouns-ch
       }
     },
     async load(api) {
-      if (api.settings.showPronouns) {
+      if (api.settings.pronouns.length > 0) {
         const services = [];
-        for (const pronounsApi of pronounsApis) {
+        for (const pronounsApi of api.settings.pronouns) {
           if (pronounsApi in knownServices) {
             services.push(
               knownServices[pronounsApi]
@@ -6802,7 +6804,7 @@ WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEM
 OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE,
  ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
-- pronouns-chat@5.0.0:
+- pronouns-chat@6.0.0:
 Licensed under MIT*.
 
 The following files have their license information within the file itself: